+++ notebooks/example.py
@@ -21,15 +21,15 @@
 # In this example, we fetch the RCMIP Radiative Forcing timeseries from `rcmip.org` and
 # produce a resource that contains a filtered subset of variables.
 
+# %%
 import logging
 import tempfile
 
-# %%
 import pooch
 import scmdata
 
 from bookshelf import LocalBook
-from bookshelf.utils import load_nb_metadata
+from bookshelf.notebook import load_nb_metadata
 
 # %% [markdown]
 # # Initialise